Rename RunDBProvider to submissiondbprovider
